@@ -5,7 +5,9 @@ import net.openwebinars.springboot.validation.error.model.impl.ApiValidationSubE
 import org.hibernate.validator.internal.engine.path.PathImpl;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.NonNull;
 import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -14,32 +16,32 @@ import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import javax.persistence.EntityNotFoundException;
-import javax.validation.ConstraintViolationException;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.ConstraintViolationException;
 import java.util.List;
 import java.util.stream.Collectors;
-
-
-
 
 @RestControllerAdvice
 public class GlobalRestControllerAdvice extends ResponseEntityExceptionHandler {
 
     @Override
-    protected ResponseEntity<Object> handleExceptionInternal(Exception ex, Object body, HttpHeaders headers, HttpStatus status, WebRequest request) {
+    protected ResponseEntity<Object> handleExceptionInternal(Exception ex,
+                                                             Object body,
+                                                             @NonNull HttpHeaders headers,
+                                                             @NonNull HttpStatusCode status,
+                                                             @NonNull WebRequest request) {
         return buildApiError(ex.getMessage(), request, status);
     }
 
-
     @ExceptionHandler({EntityNotFoundException.class})
-    public ResponseEntity<?> handleNotFoundException(EntityNotFoundException exception, WebRequest request) {
+    public ResponseEntity<?> handleNotFoundException(EntityNotFoundException exception,
+                                                     WebRequest request) {
         return buildApiError(exception.getMessage(), request, HttpStatus.NOT_FOUND);
     }
 
-
-
     @ExceptionHandler({ConstraintViolationException.class})
-    public ResponseEntity<?> handleConstraintViolationException(ConstraintViolationException exception, WebRequest request) {
+    public ResponseEntity<?> handleConstraintViolationException(ConstraintViolationException exception,
+                                                                WebRequest request) {
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
                 .body(
@@ -48,28 +50,32 @@ public class GlobalRestControllerAdvice extends ResponseEntityExceptionHandler {
                                 .message("Constraint Validation error. Please check the sublist.")
                                 .path(((ServletWebRequest) request).getRequest().getRequestURI())
                                 .subErrors(exception.getConstraintViolations().stream()
-                                        .map(v -> {
-                                            return ApiValidationSubError.builder()
-                                                    .message(v.getMessage())
-                                                    .rejectedValue(v.getInvalidValue())
-                                                    .object(v.getRootBean().getClass().getSimpleName())
-                                                    .field( ((PathImpl)v.getPropertyPath()).getLeafNode().asString())
-                                                    .build();
-                                        })
+                                        .map(v -> ApiValidationSubError.builder()
+                                                .message(v.getMessage())
+                                                .rejectedValue(v.getInvalidValue())
+                                                .object(v.getRootBean().getClass().getSimpleName())
+                                                .field( ((PathImpl)v.getPropertyPath()).getLeafNode().asString())
+                                                .build())
                                         .collect(Collectors.toList())
                                 )
                                 .build()
                 );
     }
 
-
-
     @Override
-    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
-        return buildApiErrorWithSubErrors("Validation error. Please check the sublist.", request, status, ex.getAllErrors());
+    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
+                                                                  @NonNull HttpHeaders headers,
+                                                                  @NonNull HttpStatusCode status,
+                                                                  @NonNull WebRequest request) {
+        return buildApiErrorWithSubErrors("Validation error. Please check the sublist.",
+                                            request,
+                                            status,
+                                            ex.getAllErrors());
     }
 
-    private final ResponseEntity<Object> buildApiError(String message, WebRequest request, HttpStatus status) {
+    private ResponseEntity<Object> buildApiError(String message,
+                                                 WebRequest request,
+                                                 HttpStatusCode status) {
         return ResponseEntity
                 .status(status)
                 .body(
@@ -81,7 +87,10 @@ public class GlobalRestControllerAdvice extends ResponseEntityExceptionHandler {
                 );
     }
 
-    private final ResponseEntity<Object> buildApiErrorWithSubErrors(String message, WebRequest request, HttpStatus status, List<ObjectError> subErrors) {
+    private ResponseEntity<Object> buildApiErrorWithSubErrors(String message,
+                                                              WebRequest request,
+                                                              HttpStatusCode status,
+                                                              List<ObjectError> subErrors) {
         return ResponseEntity
                 .status(status)
                 .body(
@@ -90,8 +99,8 @@ public class GlobalRestControllerAdvice extends ResponseEntityExceptionHandler {
                                 .message(message)
                                 .path(((ServletWebRequest) request).getRequest().getRequestURI())
                                 .subErrors(subErrors.stream()
-                                            .map(ApiValidationSubError::fromObjectError)
-                                            .collect(Collectors.toList())
+                                        .map(ApiValidationSubError::fromObjectError)
+                                        .collect(Collectors.toList())
                                 )
                                 .build()
                 );
