@@ -7,8 +7,8 @@ import lombok.NoArgsConstructor;
 import net.openwebinars.springboot.validation.model.Product;
 import org.hibernate.validator.constraints.URL;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotEmpty;
 
 @Data
 @NoArgsConstructor @AllArgsConstructor
