@@ -1,10 +1,9 @@
 package net.openwebinars.springboot.validation.validation.annotation;
 
-
 import net.openwebinars.springboot.validation.validation.validator.FieldsValueMatchValidator;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import java.lang.annotation.*;
 
 @Target({ ElementType.TYPE })
@@ -26,6 +25,5 @@ public @interface FieldsValueMatch {
     @interface List {
         FieldsValueMatch[] value();
     }
-
 
 }
