@@ -1,10 +1,9 @@
 package net.openwebinars.springboot.validation.validation.annotation;
 
-
 import net.openwebinars.springboot.validation.validation.validator.StrongPasswordValidator;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
 import java.lang.annotation.*;
 
 @Target({ElementType.METHOD, ElementType.FIELD})
@@ -29,6 +28,5 @@ public @interface StrongPassword {
     boolean hasNumber() default true;
 
     boolean hasSpecial() default true;
-
 
 }
