@@ -4,8 +4,8 @@ import net.openwebinars.springboot.validation.validation.annotation.PasswodsMatc
 import org.springframework.beans.PropertyAccessorFactory;
 import org.springframework.util.StringUtils;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 public class PasswordsMatchValidator implements ConstraintValidator<PasswodsMatch, Object> {
 
