@@ -3,6 +3,7 @@ package net.openwebinars.springboot.validation.error.model;
 import net.openwebinars.springboot.validation.error.model.impl.ApiErrorImpl;
 import net.openwebinars.springboot.validation.error.model.impl.ApiValidationSubError;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.validation.ObjectError;
 
 import java.time.LocalDateTime;
@@ -10,10 +11,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-
 public interface ApiError {
 
-    HttpStatus getStatus();
+    HttpStatusCode getStatus();
     int getStatusCode();
     String getMessage();
     String getPath();
@@ -23,7 +23,7 @@ public interface ApiError {
 
     static ApiError fromErrorAttributes(Map<String, Object> defaultErrorAttributesMap) {
 
-        int statusCode = ((Integer)defaultErrorAttributesMap.get("status")).intValue();
+        int statusCode = (Integer) defaultErrorAttributesMap.get("status");
 
 
         ApiErrorImpl result =
@@ -48,6 +48,5 @@ public interface ApiError {
 
         return result;
     }
-
 
 }
